dockerize

+++ database/migrations/2025_10_07_220630_remove_account_id_from_debts_table.php
@@ -12,8 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('debts', function (Blueprint $table) {
-            $table->dropForeign(['account_id']);
-            $table->dropColumn('account_id');
+            if (Schema::hasColumn('debts', 'account_id')) {
+                $table->dropForeign(['account_id']);
+                $table->dropColumn('account_id');
+            }
         });
     }
